backend/requirements.txt
pypresence
Flask
flask-cors
python-dotenv